# backend/swagger.yaml
openapi: 3.0.0
info:
  title: ThiSo API
  version: 1.0.0
  description: API documentation for ThiSo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Song:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        artist:
          type: string
        chordPro:
          type: string
    SongInput:
      type: object
      required:
        - title
        - artist
        - chordPro
      properties:
        title:
          type: string
        artist:
          type: string
        chordPro:
          type: string
    SavedSong:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        songId:
          type: string
    SavedSongInput:
      type: object
      required:
        - userId
        - songId
      properties:
        userId:
          type: string
        songId:
          type: string
paths:
...
  /songs:
    get:
      tags: [Songs]
      summary: Get all songs
      responses:
        '200':
          description: A list of songs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Song'
    post:
      tags: [Songs]
      summary: Create a new song
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SongInput'
      responses:
        '201':
          description: Song created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Song'
  /songs/{id}:
    get:
      tags: [Songs]
      summary: Get a song by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A single song
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Song'
    put:
      tags: [Songs]
      summary: Update a song by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SongInput'
      responses:
        '200':
          description: Song updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Song'
    delete:
      tags: [Songs]
      summary: Delete a song by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Song deleted
  /savedSongs:
    get:
      tags: [Songs]
      summary: Get all saved songs
      responses:
        '200':
          description: A list of saved songs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SavedSong'
    post:
      tags: [Songs]
      summary: Create a saved song
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavedSongInput'
      responses:
        '201':
          description: Saved song created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedSong'
  /savedSongs/{id}:
    get:
      tags: [Songs]
      summary: Get a saved song by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A single saved song
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedSong'
    delete:
      tags: [Songs]
      summary: Delete a saved song by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Saved song deleted